refactor: 위도, 경도 type` 변경

See merge request s09-webmobile2-sub2/S09P12B208!13

+++ square/src/main/java/com/hihi/square/domain/user/entity/CustomerAddress.java
@@ -23,18 +23,19 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name="customer_address")
+@Table(name = "customer_address")
 public class CustomerAddress {
-	@Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name="uca_id")
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "uca_id")
 	private Integer ucaId;
 
 	@ManyToOne
-	@JoinColumn(name="usr_id")
+	@JoinColumn(name = "usr_id")
 	private Customer customer;
 
 	@ManyToOne
-	@JoinColumn(name="aem_id")
+	@JoinColumn(name = "aem_id")
 	private EmdAddress emdAddress;
 
 	private String address;
@@ -42,7 +43,7 @@ public class CustomerAddress {
 	@CreatedDate
 	@Column(name = "created_at", updatable = false)
 	private LocalDateTime createdAt;
-	private Float latitude;
-	private Float longitude;
+	private Double latitude;
+	private Double longitude;
 
 }
